lib: Fi TX power ceilings
TX power ceiling have to be properly applied based on board type else
default power of 0dB will be applied dirupting the operation.

As we don't have a DTS in BM, Introduce a board type Kconfig and use it
to apply different ceilings, this works with nordic and customer boards
too.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>